pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlops-object-detection-kd"
version = "1.0.0"
description = "Hệ thống MLOps toàn diện cho Object Detection với Knowledge Distillation và TensorRT"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AI Engineer Portfolio"}
]

dependencies = [
    "ultralytics>=8.3.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "onnx>=1.15.0",
    "onnxruntime-gpu>=1.16.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.9",
    "gradio>=4.0",
    "mlflow>=2.12.0",
    "prometheus-client>=0.20.0",
    "deepchecks>=0.18.0",
    "pyyaml>=6.0",
    "pillow>=10.0",
    "numpy>=1.26.0",
    "opencv-python>=4.9.0",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
]
training = [
    "tensorboard>=2.16.0",
    "wandb>=0.17.0",
]
tensorrt = [
    "tensorrt>=10.0",
    "pycuda>=2024.1",
]
airflow = [
    "apache-airflow>=2.9.0",
    "apache-airflow-providers-cncf-kubernetes>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
